split and full divide
