+++ app/client/src/constants/AppConstants.ts
@@ -12,8 +12,8 @@ export const CANVAS_BACKGROUND_COLOR = "#FFFFFF";
 export const DEFAULT_ENTITY_EXPLORER_WIDTH = 256;
 export const DEFAULT_PROPERTY_PANE_WIDTH = 288;
 export const APP_SETTINGS_PANE_WIDTH = 525;
-export const DEFAULT_EDITOR_PANE_WIDTH = 255;
-export const DEFAULT_SPLIT_SCREEN_WIDTH = "40.4vw";
+export const DEFAULT_EXPLORER_PANE_WIDTH = 255;
+export const SPLIT_SCREEN_RATIO = 0.404;
 
 const APP_STORE_NAMESPACE = "APPSMITH_LOCAL_STORE";
