+++ app/client/src/pages/Editor/IDE/EditorPane/index.tsx
@@ -1,35 +1,29 @@
 import React from "react";
 import { Flex } from "design-system";
-import type { RouteComponentProps } from "react-router";
-import { Switch } from "react-router";
-
-import { SentryRoute } from "@appsmith/AppRouter";
-import { ADD_PATH } from "constants/routes";
-import EditorPaneSegments from "./EditorPaneSegments";
-import GlobalAdd from "./GlobalAdd";
 import { useEditorPaneWidth } from "../hooks";
-import EntityProperties from "pages/Editor/Explorer/Entity/EntityProperties";
+import EditorPaneExplorer from "./Explorer";
+import Editor from "./Editor";
+import { useSelector } from "react-redux";
+import { getIDEViewMode } from "selectors/ideSelectors";
+import { EditorViewMode } from "@appsmith/entities/IDE/constants";
 
-const EditorPane = ({ match: { path } }: RouteComponentProps) => {
+const EditorPane = () => {
   const width = useEditorPaneWidth();
+  const ideViewMode = useSelector(getIDEViewMode);
 
   return (
     <Flex
       className="ide-editor-left-pane"
-      flexDirection="column"
+      flexDirection={
+        ideViewMode === EditorViewMode.SplitScreen ? "column" : "row"
+      }
       gap="spacing-2"
       height="100%"
       overflow="hidden"
       width={width}
     >
-      {/** Entity Properties component is needed to render
-        the Bindings popover in the context menu. Will be removed eventually **/}
-      <EntityProperties />
-
-      <Switch>
-        <SentryRoute component={GlobalAdd} exact path={`${path}${ADD_PATH}`} />
-        <SentryRoute component={EditorPaneSegments} />
-      </Switch>
+      <EditorPaneExplorer />
+      <Editor />
     </Flex>
   );
 };
