+++ app/client/src/pages/Editor/IDE/hooks.ts
@@ -22,8 +22,9 @@ import {
 import { getCurrentFocusInfo } from "selectors/focusHistorySelectors";
 import { getCurrentGitBranch } from "selectors/gitSyncSelectors";
 import {
-  DEFAULT_EDITOR_PANE_WIDTH,
-  DEFAULT_SPLIT_SCREEN_WIDTH,
+  APP_SIDEBAR_WIDTH,
+  DEFAULT_EXPLORER_PANE_WIDTH,
+  SPLIT_SCREEN_RATIO,
 } from "constants/AppConstants";
 import { getIsAltFocusWidget, getWidgetSelectionBlock } from "selectors/ui";
 import { altFocusWidget, setWidgetSelectionBlock } from "actions/widgetActions";
@@ -35,6 +36,7 @@ import { FocusElement } from "navigation/FocusElements";
 import { closeJSActionTab } from "actions/jsActionActions";
 import { closeQueryActionTab } from "actions/pluginActionActions";
 import { getCurrentEntityInfo } from "../utils";
+import useWindowDimensions from "../../../utils/hooks/useWindowDimensions";
 
 export const useCurrentAppState = () => {
   const [appState, setAppState] = useState(EditorState.EDITOR);
@@ -74,7 +76,8 @@ export const useCurrentEditorState = () => {
 };
 
 export const useEditorPaneWidth = (): string => {
-  const [width, setWidth] = useState(DEFAULT_EDITOR_PANE_WIDTH + "px");
+  const [windowWidth] = useWindowDimensions();
+  const [width, setWidth] = useState(windowWidth - APP_SIDEBAR_WIDTH + "px");
   const isSideBySideEnabled = useSelector(getIsSideBySideEnabled);
   const editorMode = useSelector(getIDEViewMode);
   const { segment } = useCurrentEditorState();
@@ -86,11 +89,17 @@ export const useEditorPaneWidth = (): string => {
       segment !== EditorEntityTab.UI
     ) {
       // 1px is propertypane border width
-      setWidth(DEFAULT_SPLIT_SCREEN_WIDTH);
+      setWidth(windowWidth * SPLIT_SCREEN_RATIO + "px");
     } else {
-      setWidth(DEFAULT_EDITOR_PANE_WIDTH + "px");
+      setWidth(DEFAULT_EXPLORER_PANE_WIDTH + "px");
     }
-  }, [isSideBySideEnabled, editorMode, segment, propertyPaneWidth]);
+  }, [
+    isSideBySideEnabled,
+    editorMode,
+    segment,
+    propertyPaneWidth,
+    windowWidth,
+  ]);
 
   return width;
 };
